Raw code uploaded
This is the code as it sits in its resident folder. I don't expect you
to be able to compile it without changing a few small things.

There is little to no documentation or comments in the code, but feel
free to pm me on what things do. I'll work on making everything a bit
more formal over time, trust me. Again, I am not a programmer and this
should be evidence enough.